requirements.txt
torch>=1.7.0
torchaudio>=0.7.0
numpy>=1.19.0
librosa>=0.8.0
parselmouth>=0.4.3
webrtcvad>=2.0.10
tqdm>=4.50.0
loguru>=0.5.3
pyyaml>=5.4.1
scipy>=1.6.0
soundfile>=0.10.3
